third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/images/img-sizes-auto-expected.txt

This is a testharness.js-based test.
[FAIL] <img sizes="auto" class="test-important"> - contain-intrinsic-size
  assert_equals: expected "300px 150px" but got "30px 15px"
Harness: the test ran to completion.
